pyproject.toml
[project]
name = "videoanalyticsbot"
version = "0.1.0"
description = ""
authors = [
    {name = "Namra404",email = "armeno123@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.13,<3.15"
dependencies = [
    "mistralai (>=1.9.11,<2.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "aiogram (>=3.23.0,<4.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
